refactor: split scheduling service contracts and rotation evaluation
